requirements.txt
Pillow==9.5.0
aiogram==2.25.1
selenium==4.8.3
Flask~=2.2.3
sympy~=1.11.1